Yeasu rigs: move native command to C file, move function declration to C file. There is no point in having it in the header file if data that is used only in the C file. Export ft857_[sg]et_vfo via the ft857 header and use the header in ft897.
