+++ rigs/yaesu/frg100.h
@@ -33,36 +33,3 @@
 #define FRG100_CMD_RETCODE_OK           0x00
 #define FRG100_CMD_RETCODE_ERROR        0xF0
 
-typedef enum frg100_native_cmd_e {
-    FRG100_NATIVE_RECALL_MEM = 0,       /* 0x02, p1=ch */
-    FRG100_NATIVE_VFO_TO_MEM,           /* 0x03, p1=ch, p2=0 */
-    FRG100_NATIVE_MEM_HIDE,             /* 0x03, p1=ch, p2=1 */
-    FRG100_NATIVE_VFO_A,                /* 0x05 */
-    FRG100_NATIVE_FREQ_SET,             /* 0x0a, p1:4=freq */
-    FRG100_NATIVE_MODE_SET_LSB,         /* 0x0c, p1=0x00 */
-    FRG100_NATIVE_MODE_SET_USB,         /* 0x0c, p1=0x01 */
-    FRG100_NATIVE_MODE_SET_CW_W,        /* 0x0c, p1=0x02 */
-    FRG100_NATIVE_MODE_SET_CW_N,        /* 0x0c, p1=0x03 */
-    FRG100_NATIVE_MODE_SET_AM,          /* 0x0c, p1=0x04 */
-    FRG100_NATIVE_MODE_SET_RTTY_LSB_W,  /* 0x0c, p1=0x08 */
-    FRG100_NATIVE_MODE_SET_RTTY_USB_W,  /* 0x0c, p1=0x09 */
-    FRG100_NATIVE_MODE_SET_H3E,         /* 0x0c, p1=0x0d */
-    FRG100_NATIVE_MODE_SET_RTTY_LSB_N,  /* 0x0c, p1=0x0e */
-    FRG100_NATIVE_MODE_SET_RTTY_USB_N,  /* 0x0c, p1=0x0f */
-    FRG100_NATIVE_PTT_OFF,              /* 0x0f, p1=0 */
-    FRG100_NATIVE_PTT_ON,               /* 0x0f, p1=1 */
-    FRG100_NATIVE_UPDATE_MEM_CHNL,      /* 0x10, p1=1 */
-    FRG100_NATIVE_UPDATE_OP_DATA,       /* 0x10, p1=2 */
-    FRG100_NATIVE_UPDATE_VFO_DATA,      /* 0x10, p1=3 */
-    FRG100_NATIVE_TX_POWER_LOW,         /* 0x18 */
-    FRG100_NATIVE_TX_POWER_MID,         /* 0x28 */
-    FRG100_NATIVE_TX_POWER_HI,          /* 0x48 */
-    FRG100_NATIVE_CPY_RX_TO_TX,         /* 0x85 */
-    FRG100_NATIVE_TX_FREQ_SET,          /* 0x8a, p1:4=freq */
-    FRG100_NATIVE_OP_FREQ_STEP_UP,      /* 0x8e, p1=0 */
-    FRG100_NATIVE_OP_FREQ_STEP_DOWN,    /* 0x8e, p1=1 */
-    FRG100_NATIVE_READ_METER,           /* 0xf7 */
-    FRG100_NATIVE_READ_FLAGS,           /* 0xfa */
-    FRG100_NATIVE_SIZE
-} frg100_native_cmd_t;
-
