+++ rigs/yaesu/ft1000d.h
@@ -89,114 +89,6 @@
 
 #define FT1000D_DEFAULT_READ_TIMEOUT FT1000D_ALL_DATA_LENGTH * ( 5 + (FT1000D_PACING_INTERVAL * FT1000D_PACING_DEFAULT_VALUE))
 
-// FT1000D native commands
-typedef enum FT1000D_native_cmd_e {
-  FT1000D_NATIVE_SPLIT_OFF = 0,
-  FT1000D_NATIVE_SPLIT_ON,
-  FT1000D_NATIVE_RECALL_MEM,
-  FT1000D_NATIVE_VFO_TO_MEM,
-  FT1000D_NATIVE_LOCK_OFF,
-  FT1000D_NATIVE_LOCK_ON,
-  FT1000D_NATIVE_VFO_A,
-  FT1000D_NATIVE_VFO_B,
-  FT1000D_NATIVE_MEM_TO_VFO,
-  FT1000D_NATIVE_VFO_STEP_UP,
-  FT1000D_NATIVE_VFO_STEP_UP_FAST,
-  FT1000D_NATIVE_VFO_STEP_DOWN,
-  FT1000D_NATIVE_VFO_STEP_DOWN_FAST,
-  FT1000D_NATIVE_RX_CLARIFIER_OFF,
-  FT1000D_NATIVE_RX_CLARIFIER_ON,
-  FT1000D_NATIVE_TX_CLARIFIER_OFF,
-  FT1000D_NATIVE_TX_CLARIFIER_ON,
-  FT1000D_NATIVE_CLEAR_CLARIFIER_OFFSET,
-  FT1000D_NATIVE_CLARIFIER_OPS,
-  FT1000D_NATIVE_FREQ_SET,
-  FT1000D_NATIVE_MODE_SET_LSB,
-  FT1000D_NATIVE_MODE_SET_USB,
-  FT1000D_NATIVE_MODE_SET_CW_W,
-  FT1000D_NATIVE_MODE_SET_CW_N,
-  FT1000D_NATIVE_MODE_SET_AM_W,
-  FT1000D_NATIVE_MODE_SET_AM_N,
-  FT1000D_NATIVE_MODE_SET_FM,
-  FT1000D_NATIVE_MODE_SET_RTTY_LSB,
-  FT1000D_NATIVE_MODE_SET_RTTY_USB,
-  FT1000D_NATIVE_MODE_SET_PKT_LSB,
-  FT1000D_NATIVE_MODE_SET_PKT_FM,
-  FT1000D_NATIVE_MODE_SUB_VFOB_SET_LSB, /* Added December 2016 */
-  FT1000D_NATIVE_MODE_SUB_VFOB_SET_USB, /* Added December 2016 */
-  FT1000D_NATIVE_MODE_SUB_VFOB_SET_CW_W, /* Added December 2016 */
-  FT1000D_NATIVE_MODE_SUB_VFOB_SET_CW_N, /* Added December 2016 */
-  FT1000D_NATIVE_MODE_SUB_VFOB_SET_AM_W, /* Added December 2016 */
-  FT1000D_NATIVE_MODE_SUB_VFOB_SET_AM_N, /* Added December 2016 */
-  FT1000D_NATIVE_MODE_SUB_VFOB_SET_FM, /* Added December 2016 */
-  FT1000D_NATIVE_MODE_SUB_VFOB_SET_RTTY_LSB, /* Added December 2016 */
-  FT1000D_NATIVE_MODE_SUB_VFOB_SET_RTTY_USB, /* Added December 2016 */
-  FT1000D_NATIVE_MODE_SUB_VFOB_SET_PKT_LSB, /* Added December 2016 */
-  FT1000D_NATIVE_MODE_SUB_VFOB_SET_PKT_FM, /* Added December 2016 */
-  FT1000D_NATIVE_PACING,
-  FT1000D_NATIVE_PTT_OFF,
-  FT1000D_NATIVE_PTT_ON,
-  FT1000D_NATIVE_UPDATE_ALL_DATA,
-  FT1000D_NATIVE_UPDATE_MEM_CHNL,
-  FT1000D_NATIVE_UPDATE_OP_DATA,
-  FT1000D_NATIVE_UPDATE_VFO_DATA,
-  FT1000D_NATIVE_UPDATE_MEM_CHNL_DATA,
-  FT1000D_NATIVE_TUNER_OFF,
-  FT1000D_NATIVE_TUNER_ON,
-  FT1000D_NATIVE_TUNER_START,
-  FT1000D_NATIVE_RPTR_SHIFT_NONE,
-  FT1000D_NATIVE_RPTR_SHIFT_MINUS,
-  FT1000D_NATIVE_RPTR_SHIFT_PLUS,
-  FT1000D_NATIVE_VFO_TO_VFO,
-  FT1000D_NATIVE_SET_SUB_VFO_FREQ,
-  FT1000D_NATIVE_BANDWIDTH,
-  FT1000D_NATIVE_OP_FREQ_STEP_UP,
-  FT1000D_NATIVE_OP_FREQ_STEP_DOWN,
-  FT1000D_NATIVE_READ_METER,
-  FT1000D_NATIVE_DIM_LEVEL,
-  FT1000D_NATIVE_RPTR_OFFSET,
-  FT1000D_NATIVE_READ_FLAGS,
-  FT1000D_NATIVE_SIZE
-} FT1000D_native_cmd_t;
-
-
-/* HAMLIB API implementation */
-int ft1000d_init(RIG *rig);
-int ft1000d_cleanup(RIG *rig);
-int ft1000d_open(RIG *rig);
-int ft1000d_close(RIG *rig);
-int ft1000d_set_freq(RIG *rig, vfo_t vfo, freq_t freq);
-int ft1000d_get_freq(RIG *rig, vfo_t vfo, freq_t *freq);
-int ft1000d_set_mode(RIG *rig, vfo_t vfo, rmode_t mode, pbwidth_t width);
-int ft1000d_get_mode(RIG *rig, vfo_t vfo, rmode_t *mode, pbwidth_t *width);
-int ft1000d_set_vfo(RIG *rig, vfo_t vfo);
-int ft1000d_get_vfo(RIG *rig, vfo_t *vfo);
-int ft1000d_set_ptt(RIG *rig, vfo_t vfo, ptt_t ptt);
-int ft1000d_get_ptt(RIG *rig, vfo_t vfo, ptt_t *ptt);
-int ft1000d_set_rptr_shift(RIG *rig, vfo_t vfo, rptr_shift_t rptr_shift);
-int ft1000d_get_rptr_shift(RIG *rig, vfo_t vfo, rptr_shift_t *rptr_shift);
-int ft1000d_set_rptr_offs(RIG *rig, vfo_t vfo, shortfreq_t offs);
-int ft1000d_set_split_freq(RIG *rig, vfo_t vfo, freq_t tx_freq); /* Added December 2016 */
-int ft1000d_get_split_freq(RIG *rig, vfo_t vfo, freq_t *tx_freq); /* Added December 2016 */
-int ft1000d_set_split_vfo(RIG *rig, vfo_t vfo, split_t split, vfo_t tx_vfo);
-int ft1000d_get_split_vfo(RIG *rig, vfo_t vfo, split_t *split, vfo_t *tx_vfo);
-int ft1000d_set_split_mode(RIG *rig, vfo_t vfo, rmode_t tx_mode, pbwidth_t tx_width); /* Added December 2016 */
-int ft1000d_get_split_mode(RIG *rig, vfo_t vfo, rmode_t *tx_mode, pbwidth_t *tx_width); /* Added December 2016 */
-int ft1000d_set_rit(RIG *rig, vfo_t vfo, shortfreq_t rit);
-int ft1000d_get_rit(RIG *rig, vfo_t vfo, shortfreq_t *rit);
-int ft1000d_set_func(RIG *rig, vfo_t vfo, setting_t func, int status);
-int ft1000d_get_func(RIG *rig, vfo_t vfo, setting_t func, int *status);
-int ft1000d_set_parm(RIG *rig, setting_t parm, value_t val);
-int ft1000d_set_xit(RIG *rig, vfo_t vfo, shortfreq_t xit);
-int ft1000d_get_xit(RIG *rig, vfo_t vfo, shortfreq_t *xit);
-int ft1000d_get_level(RIG *rig, vfo_t vfo, setting_t level, value_t *value);
-int ft1000d_vfo_op(RIG *rig, vfo_t vfo, vfo_op_t op);
-int ft1000d_set_mem(RIG *rig, vfo_t vfo, int ch);
-int ft1000d_get_mem(RIG *rig, vfo_t vfo, int *ch);
-int ft1000d_set_channel (RIG *rig, vfo_t vfo, const channel_t *chan);
-int ft1000d_get_channel (RIG *rig, vfo_t vfo, channel_t *chan, int read_only);
-
-
 /*
  * The definitions below are copied from the kFT1000D
  * project and are hereby made available to the
