+++ rigs/yaesu/ft600.h
@@ -33,99 +33,4 @@
 #define FT600_WRITE_DELAY                    5
 #define FT600_POST_WRITE_DELAY               200
 #define FT600_DEFAULT_READ_TIMEOUT           2000
-
-enum ft600_native_cmd_e {
-
-  FT600_NATIVE_CAT_LOCK_ON = 0,
-  FT600_NATIVE_CAT_LOCK_OFF,
-  FT600_NATIVE_CAT_PTT_ON,
-  FT600_NATIVE_CAT_PTT_OFF,
-  FT600_NATIVE_CAT_SET_FREQ,
-  FT600_NATIVE_CAT_SET_MODE_LSB,
-  FT600_NATIVE_CAT_SET_MODE_USB,
-  FT600_NATIVE_CAT_SET_MODE_DIG,
-  FT600_NATIVE_CAT_SET_MODE_CW,
-  FT600_NATIVE_CAT_SET_MODE_AM,
-  FT600_NATIVE_CAT_CLAR_ON,
-  FT600_NATIVE_CAT_CLAR_OFF,
-  FT600_NATIVE_CAT_SET_CLAR_FREQ,
-  FT600_NATIVE_CAT_SET_VFOAB,
-  FT600_NATIVE_CAT_SET_VFOA,
-  FT600_NATIVE_CAT_SET_VFOB,
-  FT600_NATIVE_CAT_SPLIT_ON,
-  FT600_NATIVE_CAT_SPLIT_OFF,
-  FT600_NATIVE_CAT_SET_RPT_SHIFT_MINUS,
-  FT600_NATIVE_CAT_SET_RPT_SHIFT_PLUS,
-  FT600_NATIVE_CAT_SET_RPT_SHIFT_SIMPLEX,
-  FT600_NATIVE_CAT_SET_RPT_OFFSET,
-/* fix me */
-  FT600_NATIVE_CAT_SET_DCS_ON,
-  FT600_NATIVE_CAT_SET_CTCSS_ENC_ON,
-  FT600_NATIVE_CAT_SET_CTCSS_ENC_DEC_ON,
-  FT600_NATIVE_CAT_SET_CTCSS_DCS_OFF,
-/* em xif */
-  FT600_NATIVE_CAT_SET_CTCSS_FREQ,
-  FT600_NATIVE_CAT_SET_DCS_CODE,
-  FT600_NATIVE_CAT_GET_RX_STATUS,
-  FT600_NATIVE_CAT_GET_TX_STATUS,
-  FT600_NATIVE_CAT_GET_FREQ_MODE_STATUS,
-  FT600_NATIVE_CAT_PWR_WAKE,
-  FT600_NATIVE_CAT_PWR_ON,
-  FT600_NATIVE_CAT_PWR_OFF,
-  FT600_NATIVE_CAT_READ_STATUS,
-  FT600_NATIVE_CAT_READ_METERS,
-  FT600_NATIVE_CAT_READ_FLAGS
-};
-
-
-typedef enum ft600_native_cmd_e ft600_native_cmd_t;
-
-/*
- *  we are able to get way more info
- *  than we can set
- *
- */
-typedef struct
-{
-   unsigned char band_no;
-   unsigned char freq[16];
-   unsigned char mode;
-   unsigned char ctcss;
-   unsigned char dcs;
-   unsigned char flag1;
-   unsigned char flag2;
-   unsigned char clarifier[2];
-   unsigned char not_used;
-   unsigned char step1;
-   unsigned char step2;
-   unsigned char filter;
-
-   unsigned char stuffing[16];
-}
- FT600_STATUS_INFO;
-
-
-typedef struct
-{
-   unsigned char byte[8];
-}
-FT600_FLAG_INFO;
-
-
-static int ft600_init(RIG *rig);
-static int ft600_open(RIG *rig);
-static int ft600_cleanup(RIG *rig);
-static int ft600_close(RIG *rig);
-
-static int ft600_set_freq(RIG *rig, vfo_t vfo, freq_t freq);
-static int ft600_get_freq(RIG *rig, vfo_t vfo, freq_t *freq);
-
-static int ft600_set_mode(RIG *rig, vfo_t vfo, rmode_t mode, pbwidth_t width);
-static int ft600_get_mode(RIG *rig, vfo_t vfo, rmode_t *mode, pbwidth_t *width);
-
-static int ft600_get_vfo(RIG *rig, vfo_t *vfo);
-
-static int ft600_set_ptt(RIG *rig, vfo_t vfo, ptt_t ptt);
-static int ft600_get_level(RIG *rig, vfo_t vfo, setting_t level, value_t *val);
-
 #endif /* _FT600_H */
