+++ rigs/yaesu/ft840.c
@@ -41,6 +41,38 @@
 #include "ft840.h"
 
 /*
+ * Native FT840 functions. More to come :-)
+ *
+ */
+
+enum ft840_native_cmd_e {
+  FT840_NATIVE_SPLIT_OFF = 0,
+  FT840_NATIVE_SPLIT_ON,
+  FT840_NATIVE_RECALL_MEM,
+  FT840_NATIVE_VFO_TO_MEM,
+  FT840_NATIVE_VFO_A,
+  FT840_NATIVE_VFO_B,
+  FT840_NATIVE_MEM_TO_VFO,
+  FT840_NATIVE_CLARIFIER_OPS,
+  FT840_NATIVE_FREQ_SET,
+  FT840_NATIVE_MODE_SET,
+  FT840_NATIVE_PACING,
+  FT840_NATIVE_PTT_OFF,
+  FT840_NATIVE_PTT_ON,
+  FT840_NATIVE_MEM_CHNL,
+  FT840_NATIVE_OP_DATA,
+  FT840_NATIVE_VFO_DATA,
+  FT840_NATIVE_MEM_CHNL_DATA,
+  FT840_NATIVE_TUNER_OFF,
+  FT840_NATIVE_TUNER_ON,
+  FT840_NATIVE_TUNER_START,
+  FT840_NATIVE_READ_METER,
+  FT840_NATIVE_READ_FLAGS,
+  FT840_NATIVE_SIZE             /* end marker, value indicates number of */
+				                /* native cmd entries */
+};
+
+/*
  *
  * The FT-840 backend is cloned after the FT-890 backend.
  *
